feat: Moved entity/recipe pattern to separate npm-library `deep-recipes`.
The new library is included as a dependency and re-exported
from mock-objects.

Also, behavior has been changed slightly in that we no longer
do an expensive deep-clone on every set/setPath modification.
Instead we only do a spread on each level down to the end-path,
so that untouched paths are not cloned/spread.
Updates are still immutable.